bugfixes and minor refactor
